separate mod in different files
